Refactor tests for different modules, fixtures and test utils into separate modules
